+++ app/src/main/java/ru/ds/mvp_mvvm/data/api/MockLoginApiImpl.kt
@@ -1,25 +1,21 @@
-package ru.ds.mvp_mvvm.data
+package ru.ds.mvp_mvvm.data.api
 
 import ru.ds.mvp_mvvm.domain.LoginApp
 
-class WebLoginApiImpl: LoginApp {
+class MockLoginApiImpl: LoginApp {
     override fun login(login: String, password: String): Boolean {
-        Thread.sleep(2000)
-        return login ==password
+        return login == password
     }
 
     override fun register(login: String, password: String, email: String): Boolean {
-        Thread.sleep(2000)
         return login.isNotEmpty()
     }
 
     override fun logout(): Boolean {
-        Thread.sleep(2000)
-        return true
+        return  true
     }
 
     override fun forgotPassword(login: String): Boolean {
-        Thread.sleep(2000)
-        return true
+        return false
     }
 }
