+++ app/src/main/java/ru/ds/mvp_mvvm/data/api/WebLoginApiImpl.kt
@@ -1,21 +1,25 @@
-package ru.ds.mvp_mvvm.data
+package ru.ds.mvp_mvvm.data.api
 
 import ru.ds.mvp_mvvm.domain.LoginApp
 
-class MockLoginApiImpl: LoginApp {
+class WebLoginApiImpl: LoginApp {
     override fun login(login: String, password: String): Boolean {
-        return login == password
+        Thread.sleep(2000)
+        return login ==password
     }
 
     override fun register(login: String, password: String, email: String): Boolean {
+        Thread.sleep(2000)
         return login.isNotEmpty()
     }
 
     override fun logout(): Boolean {
-        return  true
+        Thread.sleep(2000)
+        return true
     }
 
     override fun forgotPassword(login: String): Boolean {
-        return false
+        Thread.sleep(2000)
+        return true
     }
 }
